refactor: extract mail list utilities
